requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2    
greenlet==3.1.1
idna==3.10
pip==24.3.1
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
typing_extensions==4.12.2
urllib3==2.2.3